microservices/schedule-ms/requirements.txt
django>=4.2,<5.0
djangorestframework
django-environ
psycopg2-binary
django-cors-headers
django-filter